Move FileSerializer to persistence packages
