+++ csb/src/main/scala/edu/msstate/dasi/csb/Benchmark.scala
@@ -5,7 +5,8 @@ import edu.msstate.dasi.csb.config.{ComponentFactory, Config}
 import edu.msstate.dasi.csb.data.distributions.DataDistributions
 import edu.msstate.dasi.csb.model.{EdgeData, VertexData}
 import edu.msstate.dasi.csb.data.seed.DataParser
-import edu.msstate.dasi.csb.util.{FileSerializer, Util}
+import edu.msstate.dasi.csb.persistence.FileSerializer
+import edu.msstate.dasi.csb.util.Util
 import edu.msstate.dasi.csb.veracity.Veracity
 import org.apache.log4j.{Level, Logger}
 import org.apache.spark.graphx.Graph
